Add fonts, copy contents of `src` to `build`
Love2D restricts access to files outside the directory with `main.lua`
in it:

```
Error: Could not open file ../res/fonts/Montserrat/
Montserrat-VariableFont_wght.ttf. Does not exist.
```

(Oh it exists all right. Was a _really_ annoying error to fix...)

We now copy the stuff to the build dir so we can access `res`.